[WIP, WORKING] Switched to 32-bit float audio and static samplerate.
Need to finish standardising samplerate conversion support and formats for XAudio and OpenAL

+++ audio/AudioMixer.cpp
@@ -24,7 +24,7 @@ namespace NovelRT::Audio
 #endif
     }
 
-    uint32_t AudioMixer::SubmitAudioBuffer(const NovelRT::Utilities::Misc::Span<int16_t> buffer, int32_t channelCount, int32_t originalSampleRate)
+    uint32_t AudioMixer::SubmitAudioBuffer(const NovelRT::Utilities::Misc::Span<float> buffer, int32_t channelCount, int32_t originalSampleRate)
     {
         auto newContext = AudioSourceContext{};
         newContext.Channels = channelCount;
